[Map] move rendering into its own package and extract out the mapAndFetch method for the markers.
